More nested components to avoid bad rendering
