+++ site.css
@@ -124,6 +124,8 @@ button:hover,
   }
 
 .progress,
+header .button.active,
+header .button.active:hover,
 .fill-purple { background-color:#6E65B3; }
 .area { fill:#6E65B3; }
 
@@ -138,12 +140,26 @@ header .fancy.title {
   line-height:60px;
   color:rgba(0,0,0,0.5);
   }
-header .button         { background-color:transparent; color:rgba(0,0,0,0.5); }
-header .button:hover   { background-color:rgba(0,0,0,0.10); color:rgba(0,0,0,0.75); }
-header .button.active  { background-color:rgba(0,0,0,0.5); color:#fff; }
+
+header .button {
+  background-color:transparent;
+  color:rgba(0,0,0,0.5);
+  }
+  header .button:hover {
+    background-color:rgba(0,0,0,0.05);
+    color:rgba(0,0,0,0.75);
+    }
+  header .button.active {
+    color:#fff;
+    }
 
 .sidebar-toggle:hover,
-.sidebar-toggle.active { background-color:rgba(0,0,0,0.10); }
+.sidebar-toggle.active {
+  background-color:rgba(0,0,0,0.10);
+  }
+.sidebar-toggle.active {
+  border-color:rgba(0,0,0,0.25);
+  }
 
 .sidebar nav a:hover {
   background-color:rgba(0,0,0,0.20);
